add partially-delayed multi-sig (#245)
* add partially-delayed multi-sig

* selectors are on a per-address basis

* refactor some code

* add tests

* add to artifacts

* timelock to uint32

* setSelectors

+++ contracts/external/multisig/DelayedMultiSig.sol
@@ -41,11 +41,11 @@ contract DelayedMultiSig is
     // ============ Events ============
 
     event ConfirmationTimeSet(uint256 indexed transactionId, uint256 confirmationTime);
-    event TimeLockChange(uint256 secondsTimeLocked);
+    event TimeLockChange(uint32 secondsTimeLocked);
 
     // ============ Storage ============
 
-    uint256 public secondsTimeLocked;
+    uint32 public secondsTimeLocked;
     mapping (uint256 => uint256) public confirmationTimes;
 
     // ============ Modifiers ============
@@ -93,7 +93,7 @@ contract DelayedMultiSig is
     constructor (
         address[] memory _owners,
         uint256 _required,
-        uint256 _secondsTimeLocked
+        uint32 _secondsTimeLocked
     )
         public
         MultiSig(_owners, _required)
@@ -110,7 +110,7 @@ contract DelayedMultiSig is
      *                             becomes executable, in seconds.
      */
     function changeTimeLock(
-        uint256 _secondsTimeLocked
+        uint32 _secondsTimeLocked
     )
         public
         onlyWallet
@@ -123,7 +123,7 @@ contract DelayedMultiSig is
 
     /**
      * Allows an owner to confirm a transaction.
-     * Overrides the function in Multisig.
+     * Overrides the function in MultiSig.
      *
      * @param  transactionId  Transaction ID.
      */
@@ -145,7 +145,7 @@ contract DelayedMultiSig is
 
     /**
      * Allows an owner to execute a confirmed transaction.
-     * Overrides the function in Multisig.
+     * Overrides the function in MultiSig.
      *
      * @param  transactionId  Transaction ID.
      */
